+++ src/frontend/jest.config.js
@@ -10,4 +10,7 @@ module.exports = {
   testMatch: [`${__dirname}/js/**/*.spec.+(ts|tsx|js)`],
   testURL: 'https://localhost',
   transformIgnorePatterns: ['node_modules/(?!(lodash-es)/)'],
+  globals: {
+    RICHIE_VERSION: 'test',
+  },
 };